Conflicts:
	hw/vigs/vigs_gl_backend_glx.c
	hw/yagl/yagl_apis/gles/yagl_gles_calls.c
	hw/yagl/yagl_backends/egl_offscreen/yagl_egl_offscreen_context.c
	hw/yagl/yagl_backends/egl_onscreen/yagl_egl_onscreen_surface.c
	hw/yagl/yagl_inc/GLES/egl.h
	hw/yagl/yagl_inc/GLES/gl.h
	hw/yagl/yagl_inc/GLES/glext.h
	hw/yagl/yagl_inc/GLES/glplatform.h
	hw/yagl/yagl_inc/GLES2/gl2.h
	hw/yagl/yagl_inc/GLES2/gl2ext.h
	hw/yagl/yagl_inc/GLES2/gl2platform.h
	hw/yagl_apis/gles/yagl_gles_api_ts.c
	hw/yagl_apis/gles/yagl_gles_api_ts.h
	hw/yagl_drivers/gles_ogl/yagl_gles_ogl.h
	hw/yagl_egl_backend.c
	hw/yagl_egl_driver.c
	hw/yagl_egl_interface.h
	hw/yagl_gles_driver.c
	hw/yagl_transport.c
	hw/yagl_transport.h
	hw/yagl_types.h
	hw/yagl_version.h

+++ tizen/src/skin/client/src/org/tizen/emulator/skin/custom/CustomScrollBar.java
@@ -1,5 +1,5 @@
 /**
- *
+ * Custom Scroll Bar
  *
  * Copyright (C) 2013 Samsung Electronics Co., Ltd. All rights reserved.
  *
@@ -189,9 +189,9 @@ public class CustomScrollBar {
 					/* draw a thumb */
 					int heightScrollGap = compositeScroll.getMinHeight() - heightScrollBar;
 
-					float tempHeightThumb = (compositeScroll.getMinHeight() - heightScrollGap) *
+					int tempHeightThumb = (compositeScroll.getMinHeight() - heightScrollGap) *
 							dataShaft.heightShaft / compositeScroll.getMinHeight();
-					dataThumb.boundsThumb.height = Math.max(1, (int)tempHeightThumb);
+					dataThumb.boundsThumb.height = Math.max(1, tempHeightThumb);
 
 					dataThumb.boundsThumb.x = 2;
 					dataThumb.boundsThumb.y = getSelection() *
@@ -328,11 +328,11 @@ public class CustomScrollBar {
 				if (dataThumb.isGrabbed == true) {
 					int yDragged = e.y - dataThumb.yGrabPosition;
 
-					float yDraggedScale = yDragged *
+					int yDraggedScale = yDragged *
 							(compositeScroll.getMinHeight() - heightScrollBar) /
 							(dataShaft.heightShaft - dataThumb.boundsThumb.height);
 
-					setSelection(dataThumb.yGrabSelection + (int)yDraggedScale);
+					setSelection(dataThumb.yGrabSelection + yDraggedScale);
 					updateScrollbar();
 				}
 			}
